+++ spec/git_pretty_accept/app_spec.rb
@@ -213,4 +213,41 @@ describe GitPrettyAccept::App do
         .to include(remote_pr_message)
     end
   end
+
+  Steps "should not merge into source branch when it is different from origin" do
+    our_change_in_master = 'our_change_in_master'
+    our_change_in_pr_branch = 'our_change_in_pr_branch'
+
+    Given 'I have a local repo' do
+      local_repo
+    end
+
+    And 'my local source branch is ahead of origin' do
+      local_repo.checkout 'master'
+      local_repo.commit_some_change our_change_in_master
+    end
+
+    And 'I have a PR_BRANCH that is not up-to-date with master' do
+      local_repo.checkout pr_branch
+      local_repo.push_some_change our_change_in_pr_branch
+    end
+
+    And 'the current branch is master' do
+      local_repo.checkout 'master'
+    end
+
+    When 'I run `git pretty-accept PR_BRANCH`' do
+      command =
+        "bundle exec #{project_path}/bin/git-pretty-accept --no-edit #{pr_branch}"
+
+      FileUtils.cd(our_path) do
+        @result = system(command, err: '/tmp/err.log')
+      end
+    end
+
+    Then 'it should not merge the PR_BRANCH due to the commit in the local source branch' do
+      expect(@result).to be_false
+      expect(File.read('/tmp/err.log')).to include('git push origin master')
+    end
+  end
 end